# Use md5.Sum()