```go
package sketch

import (
	"fmt"
	"testing"
)

func BenchmarkCount(b *testing.B) {
	s := NewSketch()
	for i := 0; i < b.N; i++ {
		ss := fmt.Sprintf("%09d", i)
		s.Add(ss)
		s.Count(ss)
	}
}
```

Before:
goos: windows
goarch: amd64
pkg: github.com/cipepser/go-sketch/sketch
BenchmarkCount-4   	  230755	      5149 ns/op
BenchmarkCount-4   	  244882	      5121 ns/op
BenchmarkCount-4   	  244882	      5174 ns/op
BenchmarkCount-4   	  235280	      5109 ns/op
BenchmarkCount-4   	  226401	      5018 ns/op
PASS
ok  	github.com/cipepser/go-sketch/sketch	6.399s

After:
goos: windows
goarch: amd64
pkg: github.com/cipepser/go-sketch/sketch
BenchmarkCount-4   	  235280	      4297 ns/op
BenchmarkCount-4   	  285697	      4365 ns/op
BenchmarkCount-4   	  279054	      4358 ns/op
BenchmarkCount-4   	  279054	      4304 ns/op
BenchmarkCount-4   	  285697	      4383 ns/op
PASS
ok  	github.com/cipepser/go-sketch/sketch	7.116s

name     old time/op  new time/op  delta
Count-4  5.11µs ± 2%  4.34µs ± 1%  -15.11%  (p=0.008 n=5+5)